Docker_Setup: Fixed the HOST_TIME_ZONE args not consumed warning (#203)
* Docker_Setup: Fixed the HOST_TIME_ZONE args not consumed warning

HOST_TIME_ZONE is passed as an argument only to ia-pybase and
ia-gobase as other modules use those as a base image. The local
machine time_zone setting logic is executed only in ia-pybase and
ia-gobase. For any other new container which doesn't use ia-gobase
and ia-pybase the time_zone setting logic must be applied and
HOST_TIME_ZONE must be passed as a build argument.

+++ VisualHmiClient/build_and_run_visualhmiclient.sh
@@ -21,12 +21,22 @@ do
     esac
 done
 
+echo "Updating container timezone..."
+# Get Docker Host timezone
+hostTimezone=`timedatectl status | grep "zone" | sed -e 's/^[ ]*Time zone: \(.*\) (.*)$/\1/g'`
+hostTimezone=`echo $hostTimezone`
+
 echo "0. Updating databus_host field in VisualHmiClient/config.json..."
 sed -i 's/"databus_host": .*/"databus_host": "'"$ETA_IP_ADDR"'",/g' VisualHmiClient/config.json
 
 echo "1. Building VisualHmiClient image..."
 cp docker_setup/dockerignores/.dockerignore .
-docker build -f VisualHmiClient/Dockerfile -t visual_hmi .
+docker build -f VisualHmiClient/Dockerfile --build-arg HOST_TIME_ZONE="$hostTimezone" -t visual_hmi .
+errorCode=`echo $?`
+if [ $errorCode != "0" ]; then
+    echo "docker build failed..."
+    exit -1
+fi
 
 echo "2. Removing any previous VisualHmiClient container..."
 docker rm -f visualhmi
